app/reminder/requirements.txt
redis>=4.5
python-dotenv>=1.0
requests>=2.31
cryptography>=41